requirements.txt
colorlog==6.10.1
homeassistant==2026.1.3
pip>=21.3.1
ruff==0.14.14
aula==0.2.0
pytest-homeassistant-custom-component
